Changed to default data types for cross-platform C

+++ lrapplications/src/pwrmanager/canconfig.h
@@ -18,8 +18,8 @@
  * Bit 23-20: Source Node ID
  * Bit 19-0: Message ID
  */
-#define MAKE_CAN_ID(srcId, dstId, msgid)		((((unsigned long)(dstId & 0x0F) << 24) | ((unsigned long)(srcId & 0x0F) << 20) | (unsigned long)(msgid & 0x7FFFF)) | CAN_EFF_FLAG)
-#define MAKE_BROADCAST_CAN_ID(srcId, msgid)		((unsigned long)(MAKE_CAN_ID(srcId, 0x00, msgid) | (((unsigned long)0x01 << 28))))
+#define MAKE_CAN_ID(srcId, dstId, msgid)		((((uint32_t)(dstId & 0x0F) << 24) | ((uint32_t)(srcId & 0x0F) << 20) | (uint32_t)(msgid & 0x7FFFF)) | CAN_EFF_FLAG)
+#define MAKE_BROADCAST_CAN_ID(srcId, msgid)		((uint32_t)(MAKE_CAN_ID(srcId, 0x00, msgid) | (((uint32_t)0x01 << 28))))
 
 #define PWRMANAGER_NODE_ID						0x01
